+++ kvtest.spec
@@ -1,5 +1,5 @@
 Name:		kvtest
-Version:	0.0.3
+Version:	0.0.4
 Release:	1%{?dist}
 License:	GPLv3
 Summary:	Key-value storages tester
@@ -45,6 +45,10 @@ Set of tools to test different file-based key-value storages.
 
 
 %changelog
+* Thu May 13 2021 TI_Eugene <ti.eugene@gmail.com> 0.0.4-1
+- Version bump
+- common as module
+
 * Wed May 09 2021 TI_Eugene <ti.eugene@gmail.com> 0.0.3-1
 - Version bump
