+++ linux.cmake
@@ -12,6 +12,7 @@ if (USE_BDB)
         add_executable(kvtest_bdb src/bdb.cpp)
         target_link_libraries(kvtest_bdb db_cxx)
         target_link_libraries(kvtest_bdb pthread)
+        target_link_libraries(kvtest_bdb common)
         install(TARGETS kvtest_bdb RUNTIME DESTINATION bin)
     else()
         set(USE_DBD OFF)
@@ -26,6 +27,7 @@ if (USE_LDB)
         add_executable(kvtest_ldb src/ldb.cpp)
         target_link_libraries(kvtest_ldb leveldb)
         target_link_libraries(kvtest_ldb pthread)
+        target_link_libraries(kvtest_ldb common)
         install(TARGETS kvtest_ldb RUNTIME DESTINATION bin)
     endif()
 endif()
@@ -38,6 +40,7 @@ if (USE_RDB)
         add_executable(kvtest_rdb src/rdb.cpp)
         target_link_libraries(kvtest_rdb PkgConfig::RDB)
         target_link_libraries(kvtest_rdb pthread)
+        target_link_libraries(kvtest_rdb common)
         install(TARGETS kvtest_rdb RUNTIME DESTINATION bin)
     else()
         set(USE_RDB OFF)
@@ -52,6 +55,7 @@ if (USE_MDB)
         add_executable(kvtest_mdb src/mdb.cpp)
         target_link_libraries(kvtest_mdb PkgConfig::MDB)
         target_link_libraries(kvtest_mdb pthread)
+        target_link_libraries(kvtest_mdb common)
         install(TARGETS kvtest_mdb RUNTIME DESTINATION bin)
     else()
         set(USE_MBD OFF)
@@ -66,6 +70,7 @@ if (USE_KC)
         add_executable(kvtest_kc src/kc.cpp)
         target_link_libraries(kvtest_kc PkgConfig::KC)
         target_link_libraries(kvtest_kc pthread)
+        target_link_libraries(kvtest_kc common)
         install(TARGETS kvtest_kc RUNTIME DESTINATION bin)
     else ()
         set(USE_KC OFF)
@@ -79,6 +84,7 @@ if (USE_TK)
         add_definitions(-DUSE_TK)
         add_executable(kvtest_tk src/tk.cpp)
         target_link_libraries(kvtest_tk PkgConfig::TK)
+        target_link_libraries(kvtest_kc common)
         install(TARGETS kvtest_tk RUNTIME DESTINATION bin)
     else ()
         set(USE_TK OFF)
