+++ macos.cmake
@@ -10,6 +10,7 @@ if (USE_BDB)
         add_executable(kvtest_bdb src/bdb.cpp)
         include_directories(/usr/local/opt/berkeley-db@4/include/)
         target_link_libraries(kvtest_bdb /usr/local/opt/berkeley-db@4/lib/libdb_cxx.dylib)
+        target_link_libraries(kvtest_bdb common)
         install(TARGETS kvtest_bdb RUNTIME DESTINATION bin)
     else()
         set(USE_BDB OFF)
@@ -24,6 +25,7 @@ if (USE_LDB)
         add_executable(kvtest_ldb src/ldb.cpp)
         include_directories(/usr/local/opt/leveldb/include/)
         target_link_libraries(kvtest_ldb /usr/local/opt/leveldb/lib/libleveldb.dylib)
+        target_link_libraries(kvtest_ldb common)
         install(TARGETS kvtest_ldb RUNTIME DESTINATION bin)
     endif()
 endif()
@@ -37,6 +39,7 @@ if (USE_RDB)
         add_executable(kvtest_rdb src/rdb.cpp)
         include_directories(/usr/local/opt/rocksdb/include/)
         target_link_libraries(kvtest_rdb /usr/local/opt/rocksdb/lib/librocksdb.dylib)
+        target_link_libraries(kvtest_rdb common)
         install(TARGETS kvtest_rdb RUNTIME DESTINATION bin)
     else()
         set(USE_RDB OFF)
@@ -52,6 +55,7 @@ if (USE_MDB)
         add_executable(kvtest_mdb src/mdb.cpp)
         include_directories(/usr/local/opt/lmdb/include/)
         target_link_libraries(kvtest_mdb /usr/local/opt/lmdb/lib/liblmdb.dylib)
+        target_link_libraries(kvtest_mdb common)
         install(TARGETS kvtest_mdb RUNTIME DESTINATION bin)
     else()
         set(USE_MDB OFF)
@@ -65,6 +69,7 @@ if (USE_KC)
         add_definitions(-DUSE_KC)
         add_executable(kvtest_kc src/kc.cpp)
         target_link_libraries(kvtest_kc PkgConfig::KC)
+        target_link_libraries(kvtest_kc common)
         install(TARGETS kvtest_kc RUNTIME DESTINATION bin)
     else()
         set(USE_KC OFF)
@@ -78,6 +83,7 @@ if (USE_TK)
         add_definitions(-DUSE_TK)
         add_executable(kvtest_tk src/tk.cpp)
         target_link_libraries(kvtest_tk PkgConfig::TK)
+        target_link_libraries(kvtest_tk common)
         install(TARGETS kvtest_tk RUNTIME DESTINATION bin)
     else ()
         set(USE_TK OFF)
