+++ src/ldb.cpp
@@ -1,8 +1,10 @@
 // LevelDB
 
 #ifdef USE_LDB
-#include "common.h"
+
+#include <iostream>
 #include <leveldb/db.h>
+#include "common.h"
 
 const filesystem::path DBNAME("kvtest.ldb");
 
@@ -22,12 +24,12 @@ bool db_open(const filesystem::path &name, bool create=false) {
  */
 bool db_sync(void) {
   if (verbose)
-    cerr << "   Sync... ";
+    std::cerr << "   Sync... ";
   time_start();
   delete db;
   auto t = time_stop();
   if (verbose)
-    cerr << t << " ms" << endl;
+    std::cerr << t << " ms" << endl;
   return true;
 }
